Refactor zfs_ioctl
Move platform specific pieces
